backend/pyproject.toml
[project]
name = "football-video-analysis-backend"
version = "0.1.0"
description = "Backend API for football video analysis using AI"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "opencv-python==4.8.1.78",
    "google-genai>=1.0.0",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "aiofiles==23.2.1",
    "numpy==1.24.3",
    "pillow==10.1.0",
    "rich==13.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

